introduce `AbstractTestCase` for functional tests
