+++ tests/Functional/UrlRewritingInControllersTest.php
@@ -12,57 +12,11 @@
 
 namespace PHPMentors\ProxyURLRewriteBundle\Functional;
 
-use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
-use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
-class UrlRewritingInControllersTest extends WebTestCase
+class UrlRewritingInControllersTest extends AbstractTestCase
 {
-    /**
-     * {@inheritdoc}
-     */
-    protected function setUp()
-    {
-        parent::setUp();
-
-        $_SERVER['KERNEL_DIR'] = __DIR__.'/app';
-        require_once $_SERVER['KERNEL_DIR'].'/AppKernel.php';
-        $_SERVER['KERNEL_CLASS'] = 'AppKernel';
-
-        $this->removeCacheDir();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function tearDown()
-    {
-        parent::tearDown();
-
-        $this->removeCacheDir();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected static function createKernel(array $options = array())
-    {
-        $kernel = KernelTestCase::createKernel($options);
-        if (array_key_exists('config', $options)) {
-            $kernel->setConfig($options['config']);
-        }
-
-        return $kernel;
-    }
-
-    protected function removeCacheDir()
-    {
-        $fileSystem = new Filesystem();
-        $fileSystem->remove($_SERVER['KERNEL_DIR'].'/cache/test');
-    }
-
     public function rewriteUrlInGenerateUrlData()
     {
         return array(
